requirements-dev.txt
# Development and Testing Dependencies

# Core dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-aiohttp>=1.0.5

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.5

# Load Testing
locust>=2.17.0

